Refactor base64 utilities
