+++ src/lib/tw-base64-utils.js
@@ -1,0 +1,19 @@
+export const base64ToArrayBuffer = base64 => {
+    const binaryString = atob(base64);
+    const len = binaryString.length;
+    const array = new Uint8Array(len);
+    for (let i = 0; i < len; i++) {
+        array[i] = binaryString.charCodeAt(i);
+    }
+    return array.buffer;
+};
+
+export const arrayBufferToBase64 = buffer => {
+    let binary = '';
+    const bytes = new Uint8Array(buffer);
+    const len = bytes.byteLength;
+    for (let i = 0; i < len; i++) {
+        binary += String.fromCharCode(bytes[i]);
+    }
+    return btoa(binary);
+};
